chore: 모듈, 패키지 구조 변경
- commerce-domain -> commerce-api
- groupId: com.example -> com.hanghae
