+++ commerce-api/src/main/kotlin/com/hanghae/commerce/HealthApplication.kt
@@ -1,0 +1,11 @@
+package com.hanghae.commerce
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class HealthApplication
+
+fun main(args: Array<String>) {
+    runApplication<HealthApplication>(*args)
+}
