+++ commerce-api/src/main/kotlin/com/hanghae/commerce/cart/application/Dummy.kt
@@ -1,0 +1,3 @@
+package com.hanghae.commerce.cart.application
+
+class Dummy
