+++ commerce-api/src/main/kotlin/com/hanghae/commerce/cart/presentaion/CartController.kt
@@ -1,0 +1,6 @@
+package com.hanghae.commerce.cart.presentaion
+
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+class CartController
