+++ commerce-api/src/main/kotlin/com/hanghae/commerce/item/application/Dummy.kt
@@ -1,0 +1,3 @@
+package com.hanghae.commerce.item.application
+
+class Dummy
