+++ commerce-api/src/main/kotlin/com/hanghae/commerce/item/presentaion/ItemController.kt
@@ -1,0 +1,6 @@
+package com.hanghae.commerce.item.presentaion
+
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+class ItemController
