+++ commerce-api/src/main/kotlin/com/hanghae/commerce/order/domain/Dummy.kt
@@ -1,0 +1,3 @@
+package com.hanghae.commerce.order.domain
+
+class Dummy
