+++ commerce-api/src/main/kotlin/com/hanghae/commerce/order/presentaion/OrderController.kt
@@ -1,0 +1,6 @@
+package com.hanghae.commerce.order.presentaion
+
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+class OrderController
