+++ commerce-api/src/main/kotlin/com/hanghae/commerce/user/application/Dummy.kt
@@ -1,0 +1,3 @@
+package com.hanghae.commerce.user.application
+
+class Dummy
