+++ commerce-api/src/main/kotlin/com/hanghae/commerce/user/domain/User.kt
@@ -1,0 +1,6 @@
+package com.hanghae.commerce.user.domain
+
+class User(
+    var id: Long? = null,
+    val name: String,
+)
