+++ commerce-api/src/main/kotlin/com/hanghae/commerce/user/domain/UserReader.kt
@@ -1,0 +1,17 @@
+package com.hanghae.commerce.user.domain
+
+import org.springframework.stereotype.Component
+
+@Component
+class UserReader(
+    private val userRepository: UserRepository,
+) {
+
+    fun save(name: String) {
+        userRepository.save(User(name = name))
+    }
+
+    fun read(id: Long): User? {
+        return userRepository.read(id)
+    }
+}
