+++ commerce-api/src/main/kotlin/com/hanghae/commerce/user/domain/UserRepository.kt
@@ -1,0 +1,6 @@
+package com.hanghae.commerce.user.domain
+
+interface UserRepository {
+    fun save(user: User)
+    fun read(id: Long): User?
+}
