+++ commerce-api/src/main/kotlin/com/hanghae/commerce/user/presentation/UserController.kt
@@ -1,0 +1,6 @@
+package com.hanghae.commerce.user.presentation
+
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+class UserController
