+++ commerce-api/src/main/resources/application-local.yml
@@ -1,0 +1,5 @@
+spring:
+
+  config:
+    activate:
+      on-profile: local
