--- commerce-domain/api/src/main/kotlin/com/example/domain/HealthApplication.kt
@@ -1,11 +1,0 @@
-package com.example.domain
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class HealthApplication
-
-fun main(args: Array<String>) {
-    runApplication<HealthApplication>(*args)
-}
