--- commerce-domain/api/src/main/kotlin/com/example/domain/cart/application/Dummy.kt
@@ -1,3 +1,0 @@
-package com.example.domain.cart.application
-
-class Dummy
