--- commerce-domain/api/src/main/kotlin/com/example/domain/cart/presentaion/CartController.kt
@@ -1,6 +1,0 @@
-package com.example.domain.cart.presentaion
-
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-class CartController
