--- commerce-domain/api/src/main/kotlin/com/example/domain/item/presentaion/ItemController.kt
@@ -1,6 +1,0 @@
-package com.example.domain.item.presentaion
-
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-class ItemController
