--- commerce-domain/api/src/main/kotlin/com/example/domain/order/domain/Dummy.kt
@@ -1,3 +1,0 @@
-package com.example.domain.order.domain
-
-class Dummy
