--- commerce-domain/api/src/main/kotlin/com/example/domain/order/presentaion/OrderController.kt
@@ -1,6 +1,0 @@
-package com.example.domain.order.presentaion
-
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-class OrderController
