--- commerce-domain/api/src/main/kotlin/com/example/domain/user/application/Dummy.kt
@@ -1,3 +1,0 @@
-package com.example.domain.user.application
-
-class Dummy
