--- commerce-domain/api/src/main/kotlin/com/example/domain/user/domain/User.kt
@@ -1,6 +1,0 @@
-package com.example.domain.user.domain
-
-class User(
-    var id: Long? = null,
-    val name: String,
-)
