--- commerce-domain/api/src/main/kotlin/com/example/domain/user/domain/UserRepository.kt
@@ -1,6 +1,0 @@
-package com.example.domain.user.domain
-
-interface UserRepository {
-    fun save(user: User)
-    fun read(id: Long): User?
-}
