--- commerce-domain/api/src/main/kotlin/com/example/domain/user/presentation/UserController.kt
@@ -1,6 +1,0 @@
-package com.example.domain.user.presentation
-
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-class UserController
