--- commerce-domain/api/src/main/resources/application-local.yml
@@ -1,5 +1,0 @@
-spring:
-
-  config:
-    activate:
-      on-profile: develop
