--- commerce-infra/db-main/src/main/kotlin/com/example/domain/data/domain/JpaUserRepository.kt
@@ -1,5 +1,0 @@
-package com.example.domain.data.domain
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface JpaUserRepository : JpaRepository<UserEntity, Long>
