--- commerce-infra/db-main/src/main/kotlin/com/example/domain/data/domain/UserEntity.kt
@@ -1,21 +1,0 @@
-package com.example.domain.data.domain
-
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.GeneratedValue
-import jakarta.persistence.GenerationType
-import jakarta.persistence.Id
-
-@Entity
-class UserEntity(
-    private val initName: String,
-) {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var id: Long? = null
-
-    @Column
-    var name: String = initName
-        protected set
-}
