+++ commerce-infra/db-main/src/main/kotlin/com/hanghae/commerce/data/domain/JpaUserRepository.kt
@@ -1,0 +1,5 @@
+package com.hanghae.commerce.data.domain
+
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface JpaUserRepository : JpaRepository<UserEntity, Long>
