+++ commerce-infra/db-main/src/main/kotlin/com/hanghae/commerce/data/domain/UserEntity.kt
@@ -1,0 +1,21 @@
+package com.hanghae.commerce.data.domain
+
+import jakarta.persistence.Column
+import jakarta.persistence.Entity
+import jakarta.persistence.GeneratedValue
+import jakarta.persistence.GenerationType
+import jakarta.persistence.Id
+
+@Entity
+class UserEntity(
+    private val initName: String,
+) {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    var id: Long? = null
+
+    @Column
+    var name: String = initName
+        protected set
+}
