+++ gradle.properties
@@ -1,5 +1,5 @@
 applicationVersion=0.0.1
-projectGroup=com.example
+projectGroup=com.hanghae
 kotlinVersion=1.8.22
 springBootVersion=3.1.3
 springDependencyManagementVersion=1.1.3
